+++ HISTORY.md
@@ -2,6 +2,79 @@
 
 ---
 
+## 1.2.2
+
+`tag:fixed` #44 #38 修改 content 内容导致高度变化应重新定位对话框。
+
+`tag:fixed` #43 多个对话框共享一个 mask 的后续 confirmbox 问题。
+
+## 1.2.1
+
+`tag:improved` 去掉了对话框的虚线框，并优化了无障碍体验。[参考](http://ued.taobao.com/blog/2011/04/onfocus-this-blur/)
+
+`tag:fixed` #40 修复一个 iframe 对话框设置高度的问题。
+
+`tag:fixed` 升级 overlay 到 1.1.3，修复遮罩层无法挡住 select 的问题。
+
+`tag:fixed` 修复同一个实例多次调用 show 后无法隐藏遮罩层的问题。（这个问题在 1.2.0 时引入）
+
+`tag:changed` #41 增加了高度超过一屏时的位置处理。
+
+
+## 1.2.0
+
+`tag:new` 新增 initialHeight 属性，用于设置 iframe 对话框的初始化高度。
+
+`tag:changed` 移除 hasMask.hideOnClick 属性。
+
+`tag:changed` 去掉对 mask 的 z-index 设置，改为调整 mask 的 dom 位置。
+
+`tag:improved` #37 支持多个对话框共享同一个 mask 的[情况](examples/two-dialog.html)。
+
+## 1.1.3
+
+`tag:improved` overlay 升级到 1.1.2，从而修复一个 IE 冗余调用 resize 的问题。
+
+`tag:changed` 修改 confirmbox 的 hasMask 默认值，使其点击遮罩层默认不关闭。
+
+## 1.1.2
+
+`tag:improved` overlay 升级到 1.1.1, mask 升级到 1.1.1, templatable 升级到 0.9.1
+
+`tag:fixed` #33 修复 firefox 下确定取消按钮文字偏下的问题。
+
+`tag:new` #34 #35 添加点击遮罩层时关闭对话框的特性，并改变了 hasMask 的默认值为 `{hideOnClick: true}`。
+
+`tag:improved` 添加 styleBox, 已解决样式冲突问题
+
+## 1.1.1
+
+`tag:fixed` #32 修复由于本地缓存导致 position 依赖打包重复的问题。
+
+## 1.1.0
+
+`tag:improved` 对 arale/overlay 的依赖从 1.0.1 升级到 1.1.0 。
+
+`tag:improved` 对 arale/mask 的依赖从 1.0.1 升级到 1.1.0 。
+
+`tag:improved` 对 arale/events 的依赖从 1.0.0 升级到 1.1.0 。
+
+`tag:improved` 对 templatable 的依赖从 arale/widget/1.0.3/templatable 升级到 arale/templatable/0.9.0/templatable 。
+
+`tag:improved` 对 handlebars 的依赖从 gallery/handlebars/1.0.0/handlebars 升级到 gallery/handlebars/1.0.2/runtime 。
+
+## 1.0.3
+
+`tag:fixed` 修复 ConfirmBox.show 的回调失效问题。
+
+## 1.0.2
+
+`tag:fixed` [#25](https://github.com/aralejs/dialog/issues/25) 使得 dialog 对 mask 的 z-index 的处理不影响其他实例。
+
+`tag:fixed` [pull/24](pull/24) 修复 ConfirmBox 静态方法生成的对象销毁失效的问题。
+
+`tag:fixed` 修复 ConfirmBox 由于 1.0.1 版本的 dist 文件打包错误导致无法使用的问题。
+
 ## 1.0.1
 
 `tag:fixed` [#21](https://github.com/aralejs/dialog/issues/21) 修复 IE6 下内容区域的高度撑不开的问题。
@@ -11,6 +84,8 @@
 `tag:changed` 去掉 close 自定义事件，可以用 before/after('hide') 来代替。
 
 `tag:improved` 对 arale/overlay 的依赖升级到 1.0.1。
+
+`tag:unresolved` 去掉 close 自定义事件导致 ConfirmBox.alert 等静态方法产生的对话框无法销毁自身。
 
 ## 1.0.0